now json parsing issues; we are close, i think
