+++ litebrite-backend/internal/domain/display.hpp
@@ -18,17 +18,25 @@ namespace domain {
     };
 
     struct Display {
-        unsigned int buffer_count;
         double fps;
         DisplayType type;
         std::optional<domain::CRGB> rgb_color;
         std::optional<domain::CRGBW> rgbw_color;
 
+        [[nodiscard]] std::string DisplayTypeToString() const {
+            if (type == DisplayType::RGB) {
+                return "RGB";
+            } else if (type == DisplayType::RGBW) {
+                return "RGBW";
+            } else if (type == DisplayType::RGB_WITH_W_INTERPOLATION) {
+                return "RGB_WITH_W_INTERPOLATION";
+            }
+        }
+
         [[nodiscard]] nlohmann::json to_json() const {
             nlohmann::json j;
-            j["buffer_count"] = buffer_count;
             j["fps"] = fps;
-            j["type"] = type;
+            j["type"] = DisplayTypeToString();
             if (type == DisplayType::RGBW) {
                 assert(rgbw_color.has_value());
                 j["rgbw_color"] = rgbw_color.value().to_json();
@@ -39,11 +47,22 @@ namespace domain {
             return j;
         }
 
+        static DisplayType DisplayTypeFromString(const std::string &display_type) {
+            if (display_type == "RGB") {
+                return DisplayType::RGB;
+            } else if (display_type == "RGBW") {
+                return DisplayType::RGBW;
+            } else if (display_type == "RGB_WITH_W_INTERPOLATION") {
+                return DisplayType::RGB_WITH_W_INTERPOLATION;
+            } else {
+                throw std::domain_error("Unknown domain::DisplayType: " + display_type);
+            }
+        }
+
         static Display from_json(const nlohmann::json& j) {
             Display d;
-            d.buffer_count = j.at("buffer_count").get<unsigned int>();
             d.fps = j.at("fps").get<double>();
-            d.type = j.at("type").get<DisplayType>();
+            d.type = DisplayTypeFromString(j.at("type").get<std::string>());
             if (d.type == DisplayType::RGBW) {
                 assert(j.contains("rgbw_color"));
                 d.rgbw_color = { CRGBW::from_json(j.at("rgbw_color")) };
