+++ litebrite-backend/internal/infrastructure/graphics/graphics.cpp
@@ -17,9 +17,9 @@ namespace infrastructure {
 
         const auto &layout = config.installation_layout;
         const auto &conf = config.installation_config;
-        const auto &display = config.display;
+        const auto &display = config.default_display;
 
-        generateBuffers(layout.universes, display.buffer_count);
+        generateBuffers(layout.universes, conf.buffer_count);
 
         if (conf.rgbw_pixels) {
             assert(conf.white_color.has_value() || conf.color_temperature.has_value());
