+++ litebrite-backend/internal/service/CMakeLists.txt
@@ -1,6 +1,6 @@
 
-set(SOURCES embedded.cpp)
-set(TARGET_LIBS pthread)
+set(SOURCES embedded_service.cpp)
+set(TARGET_LIBS art_net auth db graphics webserver)
 
 add_library(service STATIC ${SOURCES})
 target_link_libraries(service PRIVATE ${TARGET_LIBS})
